+++ src/delivery/Kontent.Ai.Delivery.Caching/GlobalUsings.cs
@@ -1,1 +1,2 @@
 global using Kontent.Ai.Delivery.Abstractions;
+global using ZiggyCreatures.Caching.Fusion.Backplane;