+++ src/delivery/Kontent.Ai.Delivery.Caching/HybridCacheManager.cs
@@ -18,14 +18,16 @@ internal sealed class HybridCacheManager(
     IDistributedCache cache,
     DeliveryCacheOptions cacheOptions,
     JsonSerializerOptions? jsonSerializerOptions = null,
-    ILogger<HybridCacheManager>? logger = null)
+    ILogger<HybridCacheManager>? logger = null,
+    IFusionCacheBackplane? backplane = null)
     : IDeliveryCacheManager, IDeliveryCachePurger, IFailSafeStateProvider, IDisposable
 {
     private readonly FusionCacheManager _inner = FusionCacheManager.CreateHybrid(
         cache,
         cacheOptions,
         jsonSerializerOptions,
-        logger);
+        logger,
+        backplane);
 
     /// <inheritdoc />
     public CacheStorageMode StorageMode => _inner.StorageMode;